Replace OpenURI with Net::HTTP
This replaces OpenURI with the slightly more verbose but less intrusive Net:HTTP interface. OpenURI monkey patches the `Kernel.open` method in a possibly unsafe way, which could be surprising behavior for a gem like this.

This also improves the performance to actually stream the response, since as far as I can tell, `io.each_line` reads the entire response first.

`request_options` have been retained as headers, but other configuration options that were supported by OpenURI  need to be considered and the docs updated. This is unlikely to be possible in a 100% backwards compatible way.